Disable building MicroG zips in favor of NanoDroid
Don't really have time to maintain zips for any unofficial forks of
MicroG, so instead I'm going to provide zips that install configuration
files for NanoDroid-MicroG so that its behavior is near-identical to
that of the zips I provide(d).

If MicroG official ever gets updated, I will continue providing zips.
Until then, this is the best way to ensure quality for users without
causing problems for myself.